+++ collector/src/test/java/com/bitmovin/analytics/utils/UtilTest.kt
@@ -1,5 +1,6 @@
 package com.bitmovin.analytics.utils;
 
+import com.bitmovin.analytics.CollectorConfig
 import org.junit.Test
 import java.lang.Exception
 import org.assertj.core.api.Assertions.*
@@ -15,5 +16,45 @@ class UtilTest {
             assertThat(top).hasSize(10)
             assertThat(top).anySatisfy{ element -> assertThat(element).contains("testTopOfStacktrace")}
         }
+    }
+
+    @Test
+    @Throws(Exception::class)
+    fun testGetIsLiveFromConfigOrPlayer_ReturnsConfigValueTrueIfPlayerNotReady() {
+        val config = CollectorConfig()
+        config.setIsLive(true)
+
+        val isLive = Util.getIsLiveFromConfigOrPlayer(false, config, false)
+        assertThat(isLive).isTrue()
+    }
+
+    @Test
+    @Throws(Exception::class)
+    fun testGetIsLiveFromConfigOrPlayer_ReturnsConfigValueFalseIfPlayerNotReady() {
+        val config = CollectorConfig()
+        config.setIsLive(false)
+
+        val isLive = Util.getIsLiveFromConfigOrPlayer(false, config, false)
+        assertThat(isLive).isFalse()
+    }
+
+    @Test
+    @Throws(Exception::class)
+    fun testGetIsLiveFromConfigOrPlayer_ReturnsPlayerIsLiveTrueIfPlayerReady() {
+        val config = CollectorConfig()
+        config.setIsLive(true)
+
+        val isLive = Util.getIsLiveFromConfigOrPlayer(true, config, true)
+        assertThat(isLive).isTrue()
+    }
+
+    @Test
+    @Throws(Exception::class)
+    fun testGetIsLiveFromConfigOrPlayer_ReturnsPlayerIsLiveFalseIfPlayerReady() {
+        val config = CollectorConfig()
+        config.setIsLive(true)
+
+        val isLive = Util.getIsLiveFromConfigOrPlayer(true, config, false)
+        assertThat(isLive).isFalse()
     }
 }